Azure_API_Ingestion_Docker/requirements.txt

requests==2.24.0
python-dotenv==0.13.0
pyodbc
sqlalchemy
